Add database creation function

+++ tests/test_pga.py
@@ -3,24 +3,27 @@ import unittest
 import os
 from sqlite3 import IntegrityError
 from lutris import pga
+from lutris.util import sql
 
 TEST_PGA_PATH = os.path.join(os.path.dirname(__file__), 'pga.db')
 
 
-class TestPersonnalGameArchive(unittest.TestCase):
+class DatabaseTester(unittest.TestCase):
     def setUp(self):
         pga.PGA_DB = TEST_PGA_PATH
         if os.path.exists(TEST_PGA_PATH):
             os.remove(TEST_PGA_PATH)
-        pga.create()
-        pga.add_game(name="LutrisTest", runner="Linux")
 
     def tearDown(self):
-        os.remove(TEST_PGA_PATH)
+        if os.path.exists(TEST_PGA_PATH):
+            os.remove(TEST_PGA_PATH)
 
-    def test_get_schema(self):
-        schema = pga.get_schema('games')
-        print schema
+
+class TestPersonnalGameArchive(DatabaseTester):
+    def setUp(self):
+        super(TestPersonnalGameArchive, self).setUp()
+        pga.create()
+        pga.add_game(name="LutrisTest", runner="Linux")
 
     def test_add_game(self):
         game_list = pga.get_games()
@@ -60,10 +63,26 @@ class TestPersonnalGameArchive(unittest.TestCase):
         self.assertEqual(game['directory'], '/foo')
 
 
-class TestDbCreator(unittest.TestCase):
+class TestDbCreator(DatabaseTester):
     def test_can_generate_fields(self):
         text_field = pga.create_field('name', 'TEXT')
         self.assertEqual(text_field, "name TEXT")
 
         id_field = pga.create_field('id', 'INTEGER', indexed=True)
         self.assertEqual(id_field, "id INTEGER PRIMARY KEY")
+
+    def test_can_create_table(self):
+        fields = [
+            {'name': 'id', 'ftype': 'INTEGER', 'indexed': True},
+            {'name': 'name', 'ftype': 'TEXT'}
+        ]
+        pga.create_table('testing', fields)
+        sql.db_insert(TEST_PGA_PATH, 'testing', {'name': "testok"})
+        results = sql.db_select(TEST_PGA_PATH, 'testing')
+        self.assertEqual(results[0]['name'], "testok")
+
+
+class TestMigration(DatabaseTester):
+    def test_get_schema(self):
+        schema = pga.get_schema('games')
+        print schema
